Конвертор валют

+++ src/ru/job4j/Convertor.java
@@ -3,19 +3,19 @@ package ru.job4j;
 public class Convertor {
     public static class Converter {
 
-        public static int rubleToEuro(int value) {
-            int rsl = value / 70;
+        public static int rubleToEuro(float value) {
+            float rsl = value / 70;
             return rsl;
         }
 
-        public static int rubleToDollar(int value) {
-            int rsl = value/60; /* формула перевода рублей в доллары. */
+        public static int rubleToDollar(float value) {
+            float rsl = value/60; /* формула перевода рублей в доллары. */
             return rsl;
         }
 
         public static void main(String[] args) {
-            int euro = Converter.rubleToEuro(140);
-            int dollor = rubleToDollar(140);
+            float euro = Converter.rubleToEuro(140);
+            float dollor = rubleToDollar(140);
             System.out.println("140 rubles are " + euro + " euro.");
             System.out.println("140 rubles are " + dollor + " dollor.");
         }
